backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
django-cors-headers==4.0.0
django-leaflet==0.28.3
psycopg2-binary==2.9.6
coreapi==2.3.3
Markdown==3.4.3
Pillow==9.5.0 